Move utils to file
